# docs/swagger.yaml
basePath: /
definitions:
  model.ErrorInput:
    properties:
      code:
        example: 400
        type: integer
      error:
        example: invalid input
        type: string
    type: object
  model.ErrorResponse:
    properties:
      code:
        example: 404
        type: integer
      error:
        example: invalid subscription ID
        type: string
    type: object
  model.ServerError:
    properties:
      code:
        example: 500
        type: integer
      error:
        example: server does not respond
        type: string
    type: object
  model.Subscription:
    properties:
      end_date:
        example: "2025-09-12T00:00:00Z"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      price:
        example: 599
        type: integer
      service_name:
        example: Yandex Plus
        type: string
      start_date:
        example: "2025-08-12T00:00:00Z"
        type: string
      user_id:
        example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        type: string
    type: object
  model.TotalCostResponse:
    properties:
      total:
        example: 1500
        type: integer
    type: object
  service.UpdateSubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: kuzmin1a.a@gmail.com
    name: Kuzmin Anton
  description: API для управления подписками пользователей
  title: Subscription Aggregator API
  version: "1.0"
paths:
  /subscriptions:
    get:
      description: Возвращает подписки с возможностью фильтрации
      parameters:
      - description: ID пользователя
        example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        in: query
        name: user_id
        type: string
      - description: Название сервиса
        example: Yandex Plus
        in: query
        name: service_name
        type: string
      - description: Начальная дата (RFC3339)
        example: "2025-01-01T00:00:00Z"
        in: query
        name: from_date
        type: string
      - description: Конечная дата (RFC3339)
        example: "2025-12-31T00:00:00Z"
        in: query
        name: to_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Subscription'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/model.ServerError'
      summary: Список подписок
      tags:
      - Subscriptions
  /subscriptions/{id}:
    delete:
      description: Удаляет подписку по ID
      parameters:
      - description: ID подписки
        example: 550e8400-e29b-41d4-a716-446655440000
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Подписка успешно удалена
        "400":
          description: Неверный ID подписки
          schema:
            $ref: '#/definitions/model.ErrorInput'
        "404":
          description: Подписка не найдена
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/model.ServerError'
      summary: Удалить подписку
      tags:
      - Subscriptions
    get:
      description: Возвращает информацию о конкретной подписке
      parameters:
      - description: ID подписки
        example: 550e8400-e29b-41d4-a716-446655440000
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Subscription'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorInput'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/model.ServerError'
      summary: Получить подписку
      tags:
      - Subscriptions
    put:
      consumes:
      - application/json
      description: Изменяет данные существующей подписки
      parameters:
      - description: ID подписки
        example: 550e8400-e29b-41d4-a716-446655440000
        in: path
        name: id
        required: true
        type: string
      - description: Новые данные подписки
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/service.UpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Подписка успешно обновлена
          schema:
            $ref: '#/definitions/model.Subscription'
        "400":
          description: Неверный формат данных
          schema:
            $ref: '#/definitions/model.ErrorInput'
        "404":
          description: Подписка не найдена
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/model.ServerError'
      summary: Обновить подписку
      tags:
      - Subscriptions
  /subscriptions/total:
    get:
      description: Возвращает общую стоимость подписок за период
      parameters:
      - description: ID пользователя
        example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        in: query
        name: user_id
        type: string
      - description: Название сервиса
        example: Yandex Plus
        in: query
        name: service_name
        type: string
      - description: Начальная дата (RFC3339)
        example: "2025-01-01T00:00:00Z"
        in: query
        name: from_date
        type: string
      - description: Конечная дата (RFC3339)
        example: "2025-12-31T00:00:00Z"
        in: query
        name: to_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TotalCostResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/model.ServerError'
      summary: Сумма подписок
      tags:
      - Subscriptions
swagger: "2.0"